tools/requirements.txt
# Pdf 文件处理相关包，占空间还挺大的...
wand
pdfminer3k
